# docs/swagger.yaml
definitions:
...
  dto.AvailabilitySlot:
    properties:
      day:
        type: string
      endTime:
        type: string
      location:
        type: string
      startTime:
        type: string
    type: object
...
  dto.CardActionRequest:
    properties:
      action:
        description: like, dislike, skip
        type: string
      targetUserId:
        type: string
    required:
    - action
    - targetUserId
    type: object
...
  dto.CreateMatchRequest:
    properties:
      availabilitySlots:
        items:
          $ref: '#/definitions/dto.AvailabilitySlot'
        type: array
      courtId:
        type: string
      matchType:
        type: string
      ntrpMax:
        type: number
      ntrpMin:
        type: number
      participantIds:
        items:
          type: string
        type: array
      playTypes:
        description: rally, singles, doubles,omitempty"`
        items:
          type: string
        type: array
      preferredCourt:
        type: string
      scheduledAt:
        type: string
      specialRequirements:
        type: string
    required:
    - matchType
    - participantIds
    type: object
...
  dto.FindMatchesRequest:
    properties:
      age_range:
        $ref: '#/definitions/services.AgeRange'
      availability:
        items:
          $ref: '#/definitions/services.AvailabilityCriteria'
        type: array
      gender:
        type: string
      limit:
        type: integer
      location:
        $ref: '#/definitions/services.LocationCriteria'
      max_distance:
        type: number
      min_reputation_score:
        type: number
      ntrp_range:
        $ref: '#/definitions/services.NTRPRange'
      play_type:
        items:
          type: string
        type: array
      playing_frequency:
        type: string
    type: object
...
  dto.UpdateReputationRequest:
    properties:
      behaviorRating:
        type: number
      matchCompleted:
        type: boolean
      wasOnTime:
        type: boolean
    type: object
...
  services.AgeRange:
    properties:
      max:
        type: integer
      min:
        type: integer
    type: object
  services.AvailabilityCriteria:
    properties:
      time:
        description: morning, afternoon, evening
        type: string
      type:
        description: weekday, weekend
        type: string
    type: object
  services.LocationCriteria:
    properties:
      city:
        type: string
      district:
        type: string
    type: object
  services.NTRPRange:
    properties:
      max:
        type: number
      min:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 網球平台後端 API 服務
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 網球平台 API
  version: "1.0"
paths:
...
  /api/v1/discovery/card-action:
    post:
      consumes:
      - application/json
      description: 處理用戶對抽卡配對的動作（喜歡、不喜歡、跳過）
      parameters:
      - description: 抽卡動作
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CardActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 處理結果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 請求錯誤
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 處理抽卡動作
      tags:
      - discovery
  /api/v1/discovery/card-history:
    get:
      description: 獲取用戶的抽卡互動歷史記錄
      parameters:
      - default: 1
        description: 頁碼
        in: query
        name: page
        type: integer
      - default: 20
        description: 每頁數量
        in: query
        name: limit
        type: integer
      - description: 動作類型篩選
        enum:
        - like
        - dislike
        - skip
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 互動歷史
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 獲取抽卡互動歷史
      tags:
      - discovery
  /api/v1/discovery/create:
    post:
      consumes:
      - application/json
      description: 創建新的球友配對
      parameters:
      - description: 配對資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 創建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 請求錯誤
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 創建配對
      tags:
      - discovery
  /api/v1/discovery/find:
    post:
      consumes:
      - application/json
      description: 根據條件探索合適的球友
      parameters:
      - description: 配對條件
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FindMatchesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 配對結果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 請求錯誤
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 尋找配對
      tags:
      - discovery
  /api/v1/discovery/history:
    get:
      description: 獲取用戶的配對歷史記錄
      parameters:
      - default: 1
        description: 頁碼
        in: query
        name: page
        type: integer
      - default: 10
        description: 每頁數量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 配對歷史
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 獲取配對歷史
      tags:
      - discovery
  /api/v1/discovery/notifications:
    get:
      description: 獲取用戶的配對相關通知
      parameters:
      - default: 1
        description: 頁碼
        in: query
        name: page
        type: integer
      - default: 20
        description: 每頁數量
        in: query
        name: limit
        type: integer
      - default: false
        description: 只顯示未讀
        in: query
        name: unread_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 通知列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 獲取配對通知
      tags:
      - discovery
  /api/v1/discovery/notifications/{notificationID}/read:
    put:
      consumes:
      - application/json
      description: 標記指定通知為已讀狀態
      parameters:
      - description: 通知ID
        in: path
        name: notificationID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 標記成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 請求錯誤
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 通知不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 標記通知為已讀
      tags:
      - discovery
  /api/v1/discovery/random:
    get:
      consumes:
      - application/json
      description: 抽卡式隨機配對功能
      parameters:
      - default: 5
        description: 配對數量
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 隨機配對結果
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 隨機配對
      tags:
      - discovery
  /api/v1/discovery/reputation:
    get:
      description: 獲取用戶的信譽分數詳情
      produces:
      - application/json
      responses:
        "200":
          description: 信譽分數
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 獲取信譽分數
      tags:
      - discovery
  /api/v1/discovery/reputation/{userID}:
    put:
      consumes:
      - application/json
      description: 更新用戶的信譽分數（通常在比賽結束後調用）
      parameters:
      - description: 用戶ID
        in: path
        name: userID
        required: true
        type: string
      - description: 信譽更新資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReputationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 請求錯誤
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 更新信譽分數
      tags:
      - discovery
  /api/v1/discovery/statistics:
    get:
      description: 獲取用戶的配對統計資訊
      produces:
      - application/json
      responses:
        "200":
          description: 統計資訊
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 伺服器錯誤
          schema:
            additionalProperties: true
            type: object
      summary: 獲取配對統計
      tags:
      - discovery
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"